requirements.txt
aiohttp==3.8.4
cassandra_driver==3.27.0
Flask==2.2.2
numpy==1.21.5
pandas==1.4.4
